似乎被我误打误撞解决了，在 dsMap.cpp 中
428，429行，添加类型转换(GLfloat)

+++ ds15gl/dsMap.cpp
@@ -425,8 +425,8 @@ void DSMap::dishi() {
         for (int j = 0; j < fenkuai * y_max - 1; j++) {
 
             glPushMatrix();
-            glTranslatef((i - fenkuai * x_max / 2) / float(fenkuai)*grid_size,
-                         (j - fenkuai * y_max / 2) / float(fenkuai)*grid_size,
+            glTranslatef((i - (GLfloat)fenkuai * x_max / 2) / float(fenkuai)*grid_size,
+                         (j - (GLfloat)fenkuai * y_max / 2) / float(fenkuai)*grid_size,
                          0);
             glBegin(GL_POLYGON);
             {
